+++ example/stairLift.py
@@ -6,4 +6,4 @@ bp_stairs.bp_stairs.render_hollow = False
 bp_stairs.make()
 stairs = bp_stairs.build()
 #show_object(stairs)
-cq.exporters.export(stairs,"stl/stairLift.stl")
+cq.exporters.export(stairs,"stl/stair_Lift.stl")