+++ example_runner.py
@@ -1,0 +1,30 @@
+
+print('control platform')
+import example.controlPlatform
+import example.controlPlatformAlt
+import example.controlPlatformPrint
+
+print('cradle')
+import example.cradle_tall
+import example.cradle
+
+print('power stations')
+import example.powerstationGreebled
+import example.powerStationGreebledUniquePanels
+import example.powerStationPlain
+import example.powerStationPowerLineMulti
+import example.powerStations_sidebyside
+
+print('spool')
+import example.spool_no_center
+import example.spool
+import example.spoolCladding
+import example.spoolCladdingGreebled
+import example.spoolLogoStack
+import example.spoolLogoStackAlt
+
+print('stair lift')
+import example.stairLift
+
+print('steel frame')
+import example.steelFrame