+++ src/cqspoolterrain/pipe/__init__.py
@@ -1,7 +1,0 @@
-from .pipeFace import pipe_face
-from .straight import straight
-from .curve import curve
-from .end import end
-from .connector import connector
-from .hatch import hatch
-from .platform import platform