+++ changes.md
@@ -2,6 +2,11 @@
 
 ## Main wip
 
+## 2.0.0
+* Move pipe code from cqspoolterrain into cqterrain
+* Move IndustrialStairs code from cqspoolterrain into cqterrain
+* Added exampe_runner.py
+
 ## 1.3.0
 * Added SpoolLogoStack and example.